requirements.txt
Django~=5.1
djangorestframework~=3.15
psycopg2-binary
black
django-filter
pytest-django
isort
flake8
djangorestframework-jsonapi
djangorestframework-jsonapi[django-filter]
djangorestframework-jsonapi[django-polymorphic]
djangorestframework-jsonapi[openapi]
pytest-cov
faker